pyproject.toml
[tool.poetry]
name = "access-key-manager"
version = "0.1.0"
description = ""
authors = ["KaySela <Kvngkhopper@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-nested-routers = "^0.94.1"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
redis = "^5.0.4"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-django = "^4.8.0"
model-bakery = "^1.18.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
